requirements.txt
fastapi==0.110.0
uvicorn==0.20.0
pydantic==2.6.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.0
shap==0.44.1
joblib==1.3.2
river==0.21.0
spacy==3.7.4
nltk==3.8.1
vaderSentiment==3.3.2
textblob==0.18.0
nrclex==3.0.0
numpy==1.26.4
pandas==2.2.0
matplotlib==3.8.2
plotly==5.19.0
scipy==1.12.0
pyyaml==6.0.1
opencv-python==4.9.0.80
deepface==0.0.89
groq==0.5.0
openai==1.12.0
google-generativeai==0.4.1
sqlalchemy==2.0.27
alembic==1.13.1
reportlab==4.1.0
qrcode[pil]==7.4.2
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.4
locust==2.23.1
Pillow==10.2.0
streamlit==1.32.0
llvmlite==0.41.1
numba==0.58.1
slowapi==0.1.9
psycopg2-binary==2.9.9
